Attempt to match the Storyboard XML better
The 'flight' command is causing large diffs to Storyboard files. We are
just adding a few tags, but the problem is that the ordering of the XML
element attributes is not alphabetical. Apple seems to have their own
order in which the attributes are written - we try to match them as well
as possible, but there are a lot of edge cases.